Move attributes pull into PLM since it’s specific to that workflow
