+++ src/workflows/plm/actions/plmProjectSpaceConfigure.ts
@@ -1,8 +1,12 @@
 import { configureSpace } from '@flatfile/plugin-space-configure';
 import * as plmBlueprints from '../blueprints/_index';
 import { FlatfileListener } from '@flatfile/listener';
+import api from '@flatfile/api';
+import { WorkbookResponse } from '@flatfile/api/api';
+import { attributes as attributesBlueprint } from '../blueprints/_index';
+import { ProductsShowApiService } from '../../../shared/products-show-api-service';
 
-export const WORKBOOK_NAME = 'PLM Import';
+const WORKBOOK_NAME = 'PLM Import';
 
 export function plmProjectSpaceConfigure(listener: FlatfileListener) {
   listener.use(
@@ -37,4 +41,57 @@ export function plmProjectSpaceConfigure(listener: FlatfileListener) {
       ],
     })
   );
+
+  // Seed the workbook with data
+  listener.on('workbook:created', async (event) => {
+    if (!event.context || !event.context.workbookId) {
+      console.error('Event context or workbookId missing');
+      return;
+    }
+
+    const workbookId = event.context.workbookId;
+    let workbook: WorkbookResponse;
+    try {
+      workbook = await api.workbooks.get(workbookId);
+    } catch (error) {
+      console.error('Error getting workbook:', error.message);
+      throw error;
+    }
+
+    const workbookName = workbook?.data?.name ? workbook.data.name : '';
+
+    if (workbookName === WORKBOOK_NAME) {
+      const sheets = workbook?.data?.sheets ? workbook.data.sheets : [];
+
+      const attributesSheet = sheets.find(
+        (s) => s.config.slug === attributesBlueprint.slug
+      );
+
+      console.log('Fetching attributes from products.show...');
+
+      // Fetch attributes from products.show API
+      const attributes = await ProductsShowApiService.fetchAttributes(event);
+
+      if (attributesSheet && attributes && attributes.length > 0) {
+        const attributeSheetId = attributesSheet.id;
+
+        const mappedAttributes = attributes.map(
+          ({ externalId, name, value, unit }) => ({
+            attribute_id: { value: externalId },
+            name: { value: name },
+            value: { value: value },
+            unit: { value: unit },
+          })
+        );
+
+        try {
+          await api.records.insert(attributeSheetId, mappedAttributes);
+        } catch (error) {
+          console.error('Error inserting attributes:', error.message);
+        }
+      }
+    } else {
+      console.error('Error - no workbook found');
+    }
+  });
 }
